pyproject.toml
[project]
name = "agentic-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-core>=0.3.64",
    "langchain-openai>=0.3.21",
    "langgraph>=0.4.8",
    "matplotlib>=3.10.3",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.5",
]
